New class for the full fitter (without d0) to be used with the AM simulation.

+++ LinearizedTrackFit/interface/BaseHistograms.h
@@ -9,14 +9,14 @@
 class BaseHistograms
 {
 public:
-  BaseHistograms(const std::string & name, const int inputSize, const int bins = 100, const float & min = -1., const float & max = 1.);
-  BaseHistograms(const std::string & name, const std::vector<std::string> & varNames, const int bins = 100, const float & min = -1., const float & max = 1.);
-  void fill(const std::vector<float> & input);
+  BaseHistograms(const std::string & name, const int inputSize, const int bins = 100, const double & min = -1., const double & max = 1.);
+  BaseHistograms(const std::string & name, const std::vector<std::string> & varNames, const int bins = 100, const double & min = -1., const double & max = 1.);
+  void fill(const std::vector<double> & input);
   void write();
 
 private:
-  void initializeVarsHistograms(const std::string name, const std::vector<std::string> & varNames, const int bins, const float & min, const float & max);
-  void initializeParsHistograms(const std::string name, const std::vector<std::string> & varNames, const int bins, const float & min, const float & max);
+  void initializeVarsHistograms(const std::string name, const std::vector<std::string> & varNames, const int bins, const double & min, const double & max);
+  void initializeParsHistograms(const std::string name, const std::vector<std::string> & varNames, const int bins, const double & min, const double & max);
   // Data members
   int inputSize_;
   std::vector<TH1F*> histograms_;
