+++ LinearizedTrackFit/src/LinearFitter.cc
@@ -9,21 +9,17 @@ LinearFitter::LinearFitter(const std::string & inputDirName) :
 
 
 // Find the matrix in the hash-map for the given geometric index (needs the input parameters)
-bool LinearFitter::fit(const std::vector<float> & vars, const float & genOneOverPt, const float & genPhi,
-    const float & genEta, const float & genZ0, const int charge, const int lastLadder)
+bool LinearFitter::fit(const std::vector<double> & vars, const double & genOneOverPt, const double & genPhi,
+                       const double & genEta, const double & genZ0, const int charge, const int lastLadder)
 {
-//  if (lastLadder != -1) geomIndex_ = lastLadder;
-  // else
   geomIndex_ = gi_(genOneOverPt, genPhi, genEta, genZ0, charge);
   return fit(vars, lastLadder);
 }
 
 
-bool LinearFitter::fit(const std::vector<float> & vars, const std::vector<StubRZPhi> & stubs,
-    const int charge, const int lastLadder)
+bool LinearFitter::fit(const std::vector<double> & vars, const std::vector<StubRZPhi> & stubs,
+                       const int charge, const int lastLadder)
 {
-//  if (lastLadder != -1) geomIndex_ = lastLadder;
-  // else
   geomIndex_ = gi_(stubs, charge);
   return fit(vars, lastLadder);
 }
@@ -65,14 +61,13 @@ void LinearFitter::readRequiredLayers(const std::string & inputFileName)
 }
 
 
-bool LinearFitter::fit(const std::vector<float> & vars, const int lastLadder)
+bool LinearFitter::fit(const std::vector<double> & vars, const int lastLadder)
 {
   if (geomIndex_ == -1) return false;
   VectorXd varsVec(vars.size());
   for (unsigned int i=0; i<vars.size(); ++i) { varsVec(i) = vars[i]; }
   if (matrices_.count(geomIndex_) == 0) {
-    matrices_.insert(std::make_pair(geomIndex_,
-        MatrixReader(inputDirName_+"matrixVD_"+std::to_string(geomIndex_)+".txt")));
+    matrices_.insert(std::make_pair(geomIndex_, MatrixReader(inputDirName_+"matrixVD_"+std::to_string(geomIndex_)+".txt")));
   }
   const auto & matrix = matrices_.find(geomIndex_)->second;
   normChi2_ = matrix.normChi2(varsVec, lastLadder);
@@ -82,7 +77,7 @@ bool LinearFitter::fit(const std::vector<float> & vars, const int lastLadder)
 
 
 // This method must be called only after the fit method to use the correct geomIndex
-std::vector<float> LinearFitter::principalComponents(const std::vector<float> & vars, const int lastLadder)
+std::vector<double> LinearFitter::principalComponents(const std::vector<double> & vars, const int lastLadder)
 {
   VectorXd varsVec(vars.size());
   for (unsigned int i=0; i<vars.size(); ++i) { varsVec(i) = vars[i]; }
@@ -91,7 +86,7 @@ std::vector<float> LinearFitter::principalComponents(const std::vector<float> & 
 
 
 // This method must be called only after the fit method to use the correct geomIndex
-std::vector<float> LinearFitter::normalizedPrincipalComponents(const std::vector<float> & vars, const int lastLadder)
+std::vector<double> LinearFitter::normalizedPrincipalComponents(const std::vector<double> & vars, const int lastLadder)
 {
   VectorXd varsVec(vars.size());
   for (unsigned int i=0; i<vars.size(); ++i) { varsVec(i) = vars[i]; }
