+++ LinearizedTrackFit/src/TreeReader.cc
@@ -127,21 +127,21 @@ bool TreeReader::nextTrack()
 }
 
 
-float TreeReader::genTrackDistanceTransverse(const float &pt, const float &phi, const float &x0, const float &y0, const int charge,
-    const float &B, const float &x1, const float &y1) const
+double TreeReader::genTrackDistanceTransverse(const double &pt, const double &phi, const double &x0, const double &y0,
+                                              const int charge, const double &B, const double &x1, const double &y1) const
 {
-  float r = pt / (0.003 * B); // In centimeters (0.3 for meters)
-  float deltaXc = x1 - (charge*r * sin(phi) + x0);
-  float deltaYc = y1 - (-charge*r * cos(phi) + y0);
+  double r = pt / (0.003 * B); // In centimeters (0.3 for meters)
+  double deltaXc = x1 - (charge*r * sin(phi) + x0);
+  double deltaYc = y1 - (-charge*r * cos(phi) + y0);
   return fabs(std::sqrt(deltaXc * deltaXc + deltaYc * deltaYc) - r);
 }
 
 
-float TreeReader::genTrackDistanceLongitudinal(const float &x0, const float &y0, const float &z0, const float &cotTheta,
-    const float &r1, const float &z1) const
+double TreeReader::genTrackDistanceLongitudinal(const double &x0, const double &y0, const double &z0, const double &cotTheta,
+    const double &r1, const double &z1) const
 {
   if (cotTheta == 0) return z1;
-  float r0 = std::sqrt(x0*x0 + y0*y0);
+  double r0 = std::sqrt(x0*x0 + y0*y0);
   // The point r, z0 is the point the track goes through, 1/cotTheta = tan(theta) = m in z = m*r + c.
   // c = z0 - m*r0
   return (z1 - z0 - (r1-r0)*cotTheta);
@@ -267,13 +267,13 @@ void TreeReader::readTrackParameters()
 }
 
 
-std::vector<float> TreeReader::getVariables()
+std::vector<double> TreeReader::getVariables()
 {
   return variables_;
 }
 
 
-std::vector<float> TreeReader::getTrackParameters()
+std::vector<double> TreeReader::getTrackParameters()
 {
   return parameters_;
 }
